rework: move answer utilities
